+++ CHANGES.md
@@ -1,10 +1,21 @@
 # Changes
 
-## [1.0.3] - unreleased
+## [1.0.4] - TBD
+
+### Changed
+
+* Upgrade `rand` dependency version to 0.7
+
+## [1.0.3] - 2019-06-28
+
+### Added
+
+* Support asynchronous data factories #850
 
 ### Changed
 
 *  Use `encoding_rs` crate instead of unmaintained `encoding` crate
+
 
 ## [1.0.2] - 2019-06-17
 
@@ -17,7 +28,7 @@
 
 ## [1.0.1] - 2019-06-17
 
-### Add
+### Added
 
 * Add support for PathConfig #903
 
@@ -42,7 +53,7 @@
 
 ## [1.0.0] - 2019-06-05
 
-### Add
+### Added
 
 * Add `Scope::configure()` method.